API

RV Data

RVModelingToolkit.RVData — Type

Stores the times, RVs, and RV errors for a single spectrograph.

Fields

t::Vector{Float64} The times of the observations in BJD.

rv::Vector{Float64} The RVs in m/s.

rverr::Vector{Float64} The RV errors in m/s.

instname::Union{String, Nothing} The name of the spectrograph.

λ::Union{Int, Nothing} The wavelength of this dataset which may be used in any modeling.

Construct an RVData object.

RVModelingToolkit.CompositeRVData — Type

Wrapper dictionary-like type to store RVs from multiple spectrographs.

RVModelingToolkit.get_times — Function

get_times(data::CompositeRVData; do_sort=true)

Generates a single vector of times of all observations, sorted in time if do_sort=true.

RVModelingToolkit.get_rvs — Function

get_rvs(data::CompositeRVData; do_sort=true)

Generates a single vector of RVs of all observations, sorted in time if do_sort=true.

RVModelingToolkit.get_rverrs — Function

get_rverrs(data::CompositeRVData; do_sort=true)

Generates a single vector of RV errors of all observations, sorted in time if do_sort=true.

RVModelingToolkit.get_instnames — Function

get_instnames(data::CompositeRVData; do_sort=true)

Generates a single vector of instrument names for all observations, sorted in time if do_sort=true.

RVModelingToolkit.get_λs — Function

get_λs(data::CompositeRVData; do_sort=true)

Generates a single vector of wavelengths of all observations, sorted in time if do_sort=true.

RVModelingToolkit.read_radvel_file — Function

read_radvel_file(fname::String, λs=nothing)

Reads a RadVel-like formatted .csv file with columns time,mnvel,errvel,tel into a CompositeRVData object.

Parameters

RVModelingToolkit.Parameter — Type

Type for a RadVel-like Bayesian Parameter.

Fields:

name::Union{String, Nothing} The name of the parameter.

value::Float64 The value of the parameter.

vary::Bool Whether or not to to fit for this parameter.

priors::Vector{Priors.Prior} A Vector of priors for this parameter.

latex_str::Union{LaTeXString, Nothing} A LaTeX-formatted string, useful for plotting.

RVModelingToolkit.Parameters — Type

Container for multiple Bayesian parameters through a dictionary-like API.

Priors

RVModelingToolkit.Priors.Prior — Type

Base types for all priors.

RVModelingToolkit.Priors.Gaussian — Type

A Gaussian prior with mean μ::Float64 and standard deviation σ::Float64.

RVModelingToolkit.Priors.Uniform — Type

A uniform prior with finite bounds.

RVModelingToolkit.Priors.JeffreysGD — Type

JeffreysGD(lower_bound::Real, upper_bound::Real, knee::Real=0.0)

Construct a JeffreysGD (~ (x-x0)^-1) bounded by lower_bound and upper_bound with a knee x0.

RVModelingToolkit.Priors.Positive — Type

A trait for a positive (improper) prior to keep a parameter > 0, useful for semi-amplitudes.

RVModelingToolkit.Priors.logprob — Function

Computes the natural logarithm of the probability for this prior given a parameter value.

The Keplerian Model

RVModelingToolkit.RVModel — Type

RVModel(planets::Union{OrderedDict{Int, OrbitBasis}, Nothing}=nothing, trend_poly_deg::Int=0; t0=2400000.0)

Construct an RVModel object to model RVs with one or multiple Keplerian orbits.

RVModelingToolkit.build — Function

build(m::RVModel, pars::Parameters, t::AbstractVector{<:Real})

Build the Keplerian + trend model, not including zero point corrections (thus only degree > 0 corrections).

RVModelingToolkit.build_planet — Function

build_planet(m::RVModel, pars::Parameters, t::AbstractVector{<:Real}, planet_index::Int)
Build the model for the single planet `planet_index`.

build_planet(t::Union{AbstractVector{<:Real}, <:Real}, per::Real, tp::Real, ecc::Real, w::Real, k::Real)

Builds the Keplerian model for a single planet on the StandardOrbitBasis.

RVModelingToolkit.build_planets — Function

build_planets(m::RVModel, pars::Parameters, t::AbstractVector{<:Real})

Build and addd all planet models together.

RVModelingToolkit.build_trend_zero — Function

build_trend_zero(m::RVModel, data::CompositeRVData, pars::Parameters, t::AbstractVector{<:Real}; instname::Union{Int, Nothing}=nothing)

Build the zero point model for each spectrograph (gamma parameters).

RVModelingToolkit.build_global_trend — Function

build_global_trend(m::RVModel, pars::Parameters, t::AbstractVector{<:Real})

Build the global trend model, not including the zero point gammas of each spectrograph (gamma dot parameters).

RVModelingToolkit.true_anomaly — Function

true_anomaly(t, tp, per, ecc)

Computes the true anomaly by solving Kepler's equation.

RVModelingToolkit.solve_kepler — Function

solve_kepler(ma::Float64, ecc::Float64)::Float64

Solves Kepler's equation with a higher order Newton method. This method is largely based on the Kepler's equation solver in RadVel.

RVModelingToolkit.tc_to_tp — Function

tc_to_tp(tc, per, ecc, w)

Converts the time of conjunction to the time of periastron.

RVModelingToolkit.tp_to_tc — Function

tp_to_tc(tp, per, ecc, w)

Converts the time of periastron to the time of conjunction.

RVModelingToolkit.disable_planet_parameters! — Function

disable_planet_parameters!(pars::Parameters, planets::OrderedDict{Int, OrbitBasis}, planet_index::Int)

Disables the parameters corresponding to planet_index.

RVModelingToolkit.solve_kepler_all_times — Function

solve_kepler_all_times(mas::AbstractVector{<:Real}, ecc::Real)

Wrapper to solve Kepler's equation for multiple observation times (multiple mean anomalies, mas) for a single value of eccentricity ecc.

Orbit Bases

RVModelingToolkit.OrbitBasis — Type

Base type for a Keplerian orbit basis through RVs.

RVModelingToolkit.StandardOrbitBasis — Type

The standard RadVel orbit basis Orbital period (per; days) Time of periastron (tp; days/BJD) Orbital eccentricity (ecc; unitless) Argument of periastron (w; radians) RV semi-amplitude (k; m/s)

RVModelingToolkit.TCOrbitBasis — Type

A common orbit basis for transiting planets. Orbital period (per; days) Time of conjunction/transit (tc; days/BJD) Orbital eccentricity (ecc; unitless) Argument of periastron (w; radians) RV semi-amplitude (k; m/s)

RVModelingToolkit.parameters — Function

parameters(b::OrbitBasis, i::Int)

Returns a vector of parameter names for this basis and planet_index.

RVModelingToolkit.convert_basis — Function

Convert from one orbit basis to another.

Gaussian Processeses

RVModelingToolkit.NoiseKernel — Type

Base type for noise kernels.

RVModelingToolkit.GaussianProcess — Type

Wrapper for a Gaussian Process noise kernel.

RVModelingToolkit.QuasiPeriodic — Type

Type for a QuasiPeriodic GP kernel. par_names is a vector containing the names of the hyper-parameters, which must be in the following order, however can use arbitrary names. 1. amplitude 2. exponential length scale 3. period length scale 4. period

RVModelingToolkit.ChromaticKernelJ1 — Type

Type for a ChromaticKernelJ1 GP kernel. par_names is a vector containing the names of the hyper-parameters, which must be in the following order, however can use arbitrary names. 1. amplitude 2. exponential length scale 3. period length scale 4. period

RVModelingToolkit.ChromaticKernelJ2 — Type

Type for a ChromaticKernelJ2 GP kernel. par_names is a vector containing the names of the hyper-parameters, which must be in the following order, however can use arbitrary names. 1. amplitude 2. exponential length scale 3. period length scale 4. period

RVModelingToolkit.compute_cov_matrix — Function

compute_cov_matrix

Method to compute the covariance matrix. Must be implemented for a noise kernel, GP, and likelihood object.

RVModelingToolkit.predict — Function

Make predictions with the GP model at arbitrary observation times.

RVModelingToolkit.compute_stationary_dist_matrix — Function

compute_stationary_dist_matrix(t1::AbstractVector{<:Real}, t2::AbstractVector{<:Real})

Computes the stationary distance matrix D[i, j] = |t[i] - t[j]|

Posteriors, Likelihoods

RVModelingToolkit.RVLikelihood — Type

RVLikelihood{M, N}

A RadVel-like likelihood object for normally distributed errors (possibly after a GP).

RVModelingToolkit.RVPosterior — Type

Construct an empty RVPosterior object

RVModelingToolkit.compute_logL — Function

compute_logL

Method to compute the logarithm of the a likelihood. Implemented by the likelihood and posterior objects.

RVModelingToolkit.compute_logaprob — Function

compute_logaprob

Method to compute the logarithm of the a posteriori probability. Implemented by the posterior object.

RVModelingToolkit.compute_prior_logprob — Function

compute_prior_logprob(post::RVPosterior, pars::Parameters)

Computes the cumulative natural logarithm of prior probability.

RVModelingToolkit.compute_residuals — Function

compute_residuals

Method to compute the residuals, still including correlated noise. Implemented by each likelihood.

RVModelingToolkit.compute_data_errors — Function

compute_data_errors

Method to compute the data errors. Implemented by each likelihood.

RVModelingToolkit.compute_noise_components — Function

compute_noise_components

Method to compute the noise components. Implemented by each likelihood / kernel.

RVModelingToolkit.compute_redχ2 — Function

compute_redχ2(residuals, errors, ν)

Utility to compute the reduced chi-squared metric.

compute_redχ2(post::RVPosterior, pars::Parameters)

RVModelingToolkit.compute_bic — Function

compute_bic(post::RVPosterior, pars::Parameters)

Computes the Bayesian information criterion (BIC).

RVModelingToolkit.compute_aicc — Function

compute_aicc(post::RVPosterior, pars::Parameters)

Computes the small-sample Akaike information criterion (AICc).

Fitting

RVModelingToolkit.run_mapfit — Function

run_mapfit(post::RVPosterior, p0::Parameters)

Perform a maximum a posteriori fit using IterativeNelderMead.jl. Returns a NamedTuple with fields: pbest::Parameters The best fit parameters. lnL::Float64 The nominal log-likelihood fcalls::Int The number of function calls. simplex::Matrix{Float64} The final simplex from the IterativeNelderMead solver. iteration::Int The final iteration of the IterativeNelderMead solver.

RVModelingToolkit.run_mcmc — Function

run_mcmc(post::RVPosterior, p0::Parameters; n_burn_steps, τrel_thresh=0.01, n_min_steps=1000, n_τs_thresh=40, check_every=100, n_max_steps=100_000)

Performs an MCMC analysis with n_burn_steps followed by a dynamic number of steps, bounded by n_min_steps and n_max_steps. Convergence is determined through measuring the median auto-correlation time (# steps). Returns a NamedTuple with fields: pbest::Parameters The set of parameters that yielded the best log-likelihood from all chains. lnLbest::Float64 The corresponding log-likelihood for the best tested parameters. pmed::Int The parameters corresponding to the 50th percentile of each distribution. lnLmed::Matrix{Float64} The corresponding log-likelihood for the median parameters. n_steps::Int The number of MCMC steps performed (ignoring burn-in). τs::Vector{Float64} The median auto-correlation times of each step. chains::Matrix{Float64} The flattened MCMC chain of shape = (n_steps, n_vary_parameters) lnLs::Vector{Float64} The corresponding log-likelihoods for the chain.

RVModelingToolkit.model_comparison — Function

model_comparison(post::RVPosterior, p0::Parameters)

Runs a MAP fit for each unique combination of planets stored in first(post)[2].model.planets. For each model, returns a NamedTuple with fields: planets::OrderedDict{Int, OrbitBasis} The planets. lnL::Float64 The nominal log likelihood. pbest::Parameters The best fit parameters. redχ2::Float64 The reduced chi square corresponding to the nominal log likelihood. bic::Float64 The Bayesian Information Criterion (BIC). aicc::Float64 The small-sample Akaike Information Criterion (AICC). n_free::Int The number of varied parameters. Δaicc::Float64 The change in the AICC relative to the model with the smallest AICC. Δbic::Float64 The change in the BIC relative to the model with the smallest AICC.

Plotting

RVModelingToolkit.plot_rvs_phased — Function

plot_rvs_phased(post::RVPosterior, pars::Parameters, planet_index::Int; data_colors::AbstractDict, titles::Bool=true, star_name::Union  {String, Nothing}=nothing)

Plot the phased RVs for a single planet. Returns a PlotlyJS figure which can be saved to the interactive HTML file fname with PlotlyJS.savefig(p, fname).

RVModelingToolkit.plot_rvs_phased_all — Function

plot_rvs_phased_all(post::RVPosterior, pars::Parameters; data_colors)

Wrapper to plot all phased RVs. Returns 2 vectors of Plotly figures, which can be saved to the interactive HTML files with PlotlyJS.savefig(p, fname).

RVModelingToolkit.plot_rvs_full — Function

plot_rvs_full(post::RVPosterior, pars::Parameters, n_model_pts=5000; data_colors=nothing, gp_colors=nothing, time_offset=nothing, gp_Δt=nothing, gp_δt=nothing)

Plots the data RVs, the Keplerian model, and the GP as a function of modified BJD, as well as the residuals (separate plots). Returns two PlotlyJS plots, which can be saved to the interactive HTML files with PlotlyJS.savefig(p, fname).

RVModelingToolkit.corner_plot — Function

corner_plot(post::RVPosterior, mcmc_result::NamedTuple)

Create a corner plot with the NamedTuple returned from run_mcmc.

Brute Force Periodogram

RVModelingToolkit.brute_force_periodogram — Function

brute_force_periodogram(post::RVPosterior, p0::Parameters, periods::AbstractVector{<:Real}, planet_index::Int; threads::Bool=true)

Generate a periodogram by performing a MAP fit for a range in fixed periods for a particular test planet. Returns a vector of NamedTuples, each with fields: - pbest::Parameters The best fit parameters. - lnL::Float64 The best fit log likelihood. - redχ2::Float64 The reduced chi-square metric. - aicc::Float64 The AICc - bic::Float64 The BIC The fits will be run in parallel using threads if threads=true (default) and Julia is running with more than one thread.

Planetary and Orbital Measurements (Coming soon)

Mass, Radius (C-K), Density, SMA ...